# Speed up brew module package install & upgrade