--- a/changelogs/fragments/9022-improve-homebrew-perf.yml
+++ b/changelogs/fragments/9022-improve-homebrew-perf.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - homebrew - speed up brew install and upgrade (https://github.com/ansible-collections/community.general/pull/9022)