@UnknownPlatypus 2024-10-19
Thank you both for the quick review! Was a pleasure to contribute!

Would you be open to a follow-up PR addressing the other optimisation ideas I left in the PR description?

Cheers!