@felixfontein 2024-10-19
There is no active module maintainer for this module, so there's nobody to say "no" :) We generally accept all PRs that look good enough and don't fail CI.